delete movingbase_node_v1
