+++ tsukuba2023/scripts/movingbase_node.py
@@ -1,136 +1,138 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 import serial
 import rospy
 from sensor_msgs.msg import Imu
 import math
 import numpy as np
+import time
 
-nowPoint=[0]*7
-HEADER = 6
-count = 0
-first_heading= 0
-def __init__(self):
-    self.port = rospy.get_param("~port", "/dev/sensors/GNSSrover")
 
-def read_relposned():
-    ackPacket=[b'\xB5',b'\x62',b'\x01',b'\x3C',b'\x00',b'\x00']
-    i = 0
-    payloadlength = 6
-    with serial.Serial("/dev/sensors/GNSSrover", 19200, timeout=1) as ser:
-        while i < payloadlength+8: 
-            incoming_char = ser.read()         
-            if (i < 3) and (incoming_char == ackPacket[i]):
-                i += 1
-            elif i == 3:
-                ackPacket[i]=incoming_char
-                i += 1           
-            elif i == 4 :
-                ackPacket[i]=incoming_char
-                i += 1
-            elif i == 5 :
-                ackPacket[i]=incoming_char        
-                payloadlength = int.from_bytes(ackPacket[4]+ackPacket[5], byteorder='little',signed=False)
-                i += 1
-            elif (i > 5):
-                ackPacket.append(incoming_char)
-                i += 1
-    if checksum(ackPacket,payloadlength) :
-        perseheading(ackPacket)
-    
-    print(ackPacket)
-    print(payloadlength)
-
-def checksum(ackPacket,payloadlength ):
-    CK_A =0
-    CK_B =0
-    for i in range(2, payloadlength+6):
-        CK_A = CK_A + int.from_bytes(ackPacket[i], byteorder='little',signed=False) 
-        CK_B = CK_B +CK_A
-    CK_A &=0xff
-    CK_B &=0xff
-    if (CK_A ==  int.from_bytes(ackPacket[-2], byteorder='little',signed=False)) and (CK_B ==  int.from_bytes(ackPacket[-1], byteorder='little',signed=False)):
-        #print("ACK Received")
-        return True
-    else :
-        print("ACK Checksum Failure:")  
-        return False
-
-def perseheading(ackPacket):
-    rospy.init_node('movingbase_yaw')
-    imu_msg = Imu()
-    pub = rospy.Publisher('/movingbase_yaw', Imu, queue_size=10)
-    print(nowPoint)
-    rate=rospy.Rate(10)
-    
-    global count
-    global first_heading
-    
-    #GPStime
-    byteoffset =4 +HEADER
-    bytevalue = ackPacket[byteoffset] 
-    for i in range(1,4):
-        bytevalue  +=  ackPacket[byteoffset+i] 
-    nowPoint[1] = int.from_bytes(bytevalue, byteorder='little',signed=True)
-    nowPoint[2] =nowPoint[1]/1000
-    print("GPStime:%f sec" %float(nowPoint[2]))
-    
-    #Carrier solution status
-    flags = int.from_bytes(ackPacket[60 + HEADER], byteorder='little',signed=True)
-    nowPoint[3] =  flags  & (1 << 0) #gnssFixOK 
-    nowPoint[4] =  (flags   & (0b11 <<3)) >> 3 #carrSoln0:no carrier 1:float 2:fix
-    print("gnssFixOk:%d" %nowPoint[3])
-    print("carrSoln:%d" %nowPoint[4])
-    
-    #relPosHeading
-    byteoffset =24 +HEADER
-    bytevalue = ackPacket[byteoffset] 
-    for i in range(1,4):
-        bytevalue  +=  ackPacket[byteoffset+i] 
-    nowPoint[5] = int.from_bytes(bytevalue, byteorder='little',signed=True) 
-    #print("heading:%f deg" %float(nowPoint[5]/100000))
-    print(nowPoint)
-    #0~360
-    heading=nowPoint[5]/100000+90
-    if heading >= 360: heading -= 360
-    
-    if(count == 0):
-        first_heading=heading   
-        count = 1
+class movingbase_yaw:
+    def __init__(self):
+        self.HEADER = 6
+        self.count = 0
+        self.first_heading = 0             
+        self.port = rospy.get_param("~port", "/dev/sensors/GNSSrover")
+        self.baudrate = rospy.get_param("~baud", 115200)#19200
+        self.time_out = rospy.get_param("~time_out", 0.5)#1
         
-    relative_heading=heading-first_heading    
-    if relative_heading < 0:
-        relative_heading += 360
+        self.imu_msg = Imu()
+        self.pub = rospy.Publisher('/movingbase_yaw', Imu, queue_size=10)
     
-    #-180~-1
-    if relative_heading>180:
-        relative_heading -= 360
-    
-    
-    movingbaseyaw=relative_heading*(math.pi/180)#deg>radian   
+    def readrelposned(self):
+        ackPacket=[b'\xB5',b'\x62',b'\x01',b'\x3C',b'\x00',b'\x00']
+        i = 0
+        payloadlength = 6    
+        with serial.Serial(self.port, self.baudrate, timeout=self.time_out) as ser:
+            while i < payloadlength+8: 
+                incoming_char = ser.read()         
+                if (i < 3) and (incoming_char == ackPacket[i]):
+                    i += 1
+                elif i == 3:
+                    ackPacket[i]=incoming_char
+                    i += 1           
+                elif i == 4 :
+                    ackPacket[i]=incoming_char
+                    i += 1
+                elif i == 5 :
+                    ackPacket[i]=incoming_char        
+                    payloadlength = int.from_bytes(ackPacket[4]+ackPacket[5], byteorder='little',signed=False)
+                    i += 1
+                elif (i > 5):
+                    ackPacket.append(incoming_char)
+                    i += 1
+        
+        if self.checksum(ackPacket, payloadlength) :#1
+            print("checksum ok")
+            nowpoint_info = self.perseheading(ackPacket)#2           
+            return nowpoint_info
 
-    #print("robotheading:%f deg" %float(nowPoint[6]))
-    #print("firstheading:%f rad" %float(first_heading))
-    #print("robotheading:%f deg" %float(absolute_heading))
-    #print("firstheading:%f rad" %float(relative_heading))
-    
-    #print("yaw:%f radian"%float(movingbaseyaw)) 
-    #print("yaw:%f w"%float(np.sin(movingbaseyaw))) 
-    #print("yaw:%f w"%float(np.sin(movingbaseyaw))) 
-    
-    imu_msg.header.stamp = rospy.get_rostime()
-    
-    imu_msg.orientation.x = 0
-    imu_msg.orientation.y = 0
-    imu_msg.orientation.z = movingbaseyaw #not orientation.z>>yaw
-    imu_msg.orientation.w = 0
-    
-    if nowPoint[4] == 2:
-        pub.publish(imu_msg)
-    
-    rate.sleep()
-    #rospy.spin()
-    return nowPoint
+    def checksum(self, ackPacket, payloadlength):#1
+        CK_A = 0
+        CK_B = 0
+        for i in range(2, payloadlength+6):
+            CK_A = CK_A + int.from_bytes(ackPacket[i], byteorder='little',signed=False) 
+            CK_B = CK_B + CK_A
+        CK_A &= 0xff
+        CK_B &= 0xff
+        if (CK_A ==  int.from_bytes(ackPacket[-2], byteorder = 'little',signed=False)) and (CK_B ==  int.from_bytes(ackPacket[-1], byteorder='little',signed=False)):
+            #print("ACK Received")
+            return True
+        else :
+            print("ACK Checksum Failure:")  
+            return False
 
-while 1:
+    def perseheading(self,ackPacket):#2
+        nowPoint = []
+        #GPStime
+        byteoffset = 4 +self.HEADER
+        bytevalue = ackPacket[byteoffset] 
+        for i in range(1,4):
+            bytevalue  +=  ackPacket[byteoffset+i] 
+        time = int.from_bytes(bytevalue, byteorder='little',signed=True)
+        gpstime = time/1000
+        #print("GPStime:%f sec" %float(gpstime))
+        nowPoint.append(gpstime)#0
+        
+        #Carrier solution status
+        flags = int.from_bytes(ackPacket[60 + self.HEADER], byteorder='little',signed=True)
+        gnssFixOK  =  flags  & (1 << 0) #gnssFixOK 
+        carrSoln =  (flags   & (0b11 <<3)) >> 3 #carrSoln0:no carrier 1:float 2:fix
+        #print("gnssFixOk:%d" %gnssFixOK)
+        #print("carrSoln:%d" %carrSoln)
+        nowPoint.append(gnssFixOK)#1
+        nowPoint.append(carrSoln)#2
+        
+        #relPosHeading
+        byteoffset = 24 +self.HEADER
+        bytevalue = ackPacket[byteoffset] 
+        for i in range(1,4):
+            bytevalue  +=  ackPacket[byteoffset+i] 
+        heading = int.from_bytes(bytevalue, byteorder='little',signed=True) 
+        #print("heading:%f deg" %float(heading/100000))
+        nowPoint.append(heading/100000)#3
+        
+        return nowPoint
+        
+    def movingbase_pub(self):
+        nowpoint = self.readrelposned()
+        
+        #0~360
+        heading = nowpoint[3]+90
+        if heading >= 360: 
+            heading -= 360
+        
+        if self.count == 0:
+            self.first_heading=heading   
+            self.count = 1
+        
+        relative_heading = heading - self.first_heading    
+        if relative_heading < 0:
+            relative_heading += 360
+    
+        #-180~-1
+        if relative_heading>180:
+            relative_heading -= 360
+        
+        movingbaseyaw=relative_heading*(math.pi/180)#deg>radian   
+
+        #print("yaw:%f w"%float(np.sin(movingbaseyaw))) 
+    
+        self.imu_msg.header.stamp = rospy.Time.now()
+    
+        self.imu_msg.orientation.x = heading#0~360 deg
+        self.imu_msg.orientation.y = 0
+        self.imu_msg.orientation.z = movingbaseyaw #not orientation.z>>yaw
+        self.imu_msg.orientation.w = 0
+    
+        if nowpoint[2] == 2:
+            self.pub.publish(self.imu_msg)
+           
+        
+if __name__ == "__main__":
+    rospy.init_node('movingbase_yaw')
+    moving_base_yaw = movingbase_yaw()
+    rate = rospy.Rate(10)
     while not rospy.is_shutdown():
-        read_relposned()
+        moving_base_yaw.movingbase_pub()
+        rate.sleep()
